Reorder headers, move ScheduleMarkerPublisher into separate file, other feedback
