+++ rmf_schedule_visualizer/package.xml
@@ -18,7 +18,7 @@
   <depend>nav_msgs</depend>
   <depend>visualization_msgs</depend>
   <depend>rmf_schedule_visualizer_msgs</depend>
-  <depend>building_map_msgs</depend>
+  <depend>rmf_building_map_msgs</depend>
   <depend>boost</depend>
   <depend>libboost-date-time-dev</depend>
   <depend>libboost-regex-dev</depend>
